+++ src/main/java/com/recipe/gpt/app/domain/recipe/procedure/ProcedureItem.java
@@ -1,12 +1,13 @@
-package com.recipe.gpt.app.domain.board.ingredient;
+package com.recipe.gpt.app.domain.recipe.procedure;
 
-import com.recipe.gpt.app.domain.board.Board;
+import com.recipe.gpt.app.domain.recipe.Recipe;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -16,27 +17,24 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class IngredientItem {
+public class ProcedureItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 20)
-    private String name;
-
-    @Column(nullable = false, length = 10)
-    private String quantity;
+    @Lob
+    @Column(nullable = false, length = 200)
+    private String description;
 
     @ManyToOne
-    @JoinColumn(name = "board_id")
-    private Board board;
+    @JoinColumn(name = "recipe_id")
+    private Recipe recipe;
 
     @Builder
-    private IngredientItem(Board board, String name, String quantity) {
-        this.board = board;
-        this.name = name;
-        this.quantity = quantity;
+    private ProcedureItem(Recipe recipe, String description) {
+        this.recipe = recipe;
+        this.description = description;
     }
 
 }
